busca de todos os times, feita a funcao para pegar os dados de todos os jogadores daquele time e corrigido a quantidade de vezes do campeao mais jogado de cada jogador
