--- Python/cartinhasLoud.py
@@ -1,81 +1,0 @@
-import requests
-from bs4 import BeautifulSoup
-import pandas as pd
-import matplotlib.pyplot as plt
-import re
-import json
-
-def extract_champion_data(champion_info):
-    champion = ''
-    winrate = ''
-    kda = ''
-    
-    match_champion = re.search(r'(.+?)\s+Winrate', champion_info)
-    if match_champion:
-        champion = match_champion.group(1).strip()
-    
-    match_winrate = re.search(r'Winrate\s*:\s*([\d.]+%?)', champion_info)
-    if match_winrate:
-        winrate = match_winrate.group(1).strip()
-    
-    match_kda = re.search(r'KDA\s*:\s*([\d.]+)', champion_info)
-    if match_kda:
-        kda = match_kda.group(1).strip()
-    
-    match_games = re.search(r'(\d+)\s*$', champion_info, re.MULTILINE)
-    if match_games:
-        games = match_games.group(1).strip()
-    
-    return {
-        'Champion': champion,
-        'Winrate': winrate,
-        'KDA': kda,
-    }
-
-url = "https://gol.gg/teams/team-stats/2180/split-ALL/tournament-CBLOL%20Split%201%202024/"
-headers = {
-    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3'}
-
-response = requests.get(url, headers=headers)
-
-if response.status_code == 200:
-    soup = BeautifulSoup(response.text, 'html.parser')
-
-    table = soup.find('table', class_='footable')
-
-    if table:
-        data = []
-        for row in table.find_all('tr'):
-            columns = row.find_all(['td', 'th'])
-            data.append([column.text.strip() for column in columns])
-
-        for row in data[1:]:
-            if len(row) == 8:
-                champion_info = row[-1]
-                champion_data = extract_champion_data(champion_info)
-                row[-1] = champion_data
-        print(data)
-        
-        json_data = json.dumps(data)
-        
-        with open('cartinhas.json', 'w') as f:
-            f.write(json_data)
-        
-        df = pd.DataFrame(data[1:], columns=data[0])
-        
-        fig, ax = plt.subplots(figsize=(10, 6))
-        ax.axis('off')
-
-        tab = ax.table(cellText=df.values, colLabels=df.columns, cellLoc='center', loc='center', colColours=['#f0f0f0']*len(df.columns))
-        tab.auto_set_font_size(False)
-        tab.set_fontsize(10)
-
-        tab.auto_set_column_width([i for i in range(len(df.columns))])
-
-        plt.savefig('tabelaLoud.png', bbox_inches='tight', pad_inches=0.5)
-        # plt.show()
-        
-    else:
-        print("Nenhuma tabela encontrada no site.")
-else:
-    print('Erro ao acessar o site:', response.status_code)
