+++ cartinhas.json
@@ -1,1 +1,1 @@
-[["Role", "Player", "KDA", "KP %", "VSPM", "Dmg %", "Gold %", "Champions played"], ["TOP", "Robo", "2.3", "55.2%", "1.2", "25.4%", "22.3%", {"Champion": "Jax", "Winrate": "100%", "KDA": "3.4"}], ["JUNGLE", "Croc", "3.2", "73.3%", "1.46", "12.6%", "17.7%", {"Champion": "Sejuani", "Winrate": "75%", "KDA": "5.5"}], ["MID", "Tinowns", "9.6", "70.2%", "1.33", "28.3%", "22.8%", {"Champion": "Orianna", "Winrate": "50%", "KDA": "7.0"}], ["BOT", "Route", "5.6", "66.4%", "2.04", "26.2%", "22.8%", {"Champion": "Kalista", "Winrate": "50%", "KDA": "3.3"}], ["SUPPORT", "RedBert", "3.3", "77.6%", "2.98", "7.6%", "14.4%", {"Champion": "Nautilus", "Winrate": "50%", "KDA": "2.2"}]]
+[["Role", "Player", "KDA", "KP %", "VSPM", "Dmg %", "Gold %", "Champions played"], ["Last line-up"], ["TOP", "Guigo", "4.6", "62%", "1.11", "27.1%", "23.4%", {"Champion": "Aatrox", "Winrate": "80%", "KDA": "5.8", "Times Played": "5"}], ["JUNGLE", "Disamis", "5.1", "71.1%", "1.42", "12.3%", "17.4%", {"Champion": "Rell", "Winrate": "80%", "KDA": "6.4", "Times Played": "5"}], ["MID", "Leleko", "7.6", "71.4%", "1.25", "23.5%", "20.9%", {"Champion": "Orianna", "Winrate": "100%", "KDA": "14.3", "Times Played": "4"}], ["BOT", "SMILEY", "5.6", "67.3%", "1.41", "25.6%", "23.7%", {"Champion": "Kalista", "Winrate": "40%", "KDA": "2.6", "Times Played": "5"}], ["SUPPORT", "ProDelta", "5", "76.3%", "3.81", "8.4%", "14.2%", {"Champion": "Nautilus", "Winrate": "66.7%", "KDA": "2.5", "Times Played": "3"}], ["Subs"], ["MID", "Toucouille", "3.7", "62.7%", "0.91", "30.3%", "21.8%", {"Champion": "Azir", "Winrate": "75%", "KDA": "3.5", "Times Played": "4"}]]
