+++ css/style.css
@@ -1,0 +1,7 @@
+body {
+    background-color: #0d1117;
+}
+
+.tabelasTimes img {
+    width: 80%;
+}
