+++ index.html
@@ -1,13 +1,53 @@
-<!-- index.html -->
 <!DOCTYPE html>
-<html lang="en">
+<html lang="pt-br">
+
 <head>
     <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+    <link rel="stylesheet" href="css/style.css">
     <title>Cartinhas CBLOL</title>
 </head>
+
 <body>
-    <div id="conteudo"></div>
+    <div class="container w-100">
+        <div class="col">
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela Fluxo.png" alt="tabela fluxo" >
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela FURIA Esports.png" alt="tabela furia">
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela INTZ e-Sports.png" alt="tabela intz">
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela KaBuM! e-Sports.png" alt="tabela kabum">
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela Liberty.png" alt="tabela liberty">
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela Los Grandes.png" alt="tabela los">
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela LOUD.png" alt="tabela loud">
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela paiN Gaming.png" alt="tabela pain">
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela RED Canids.png" alt="tabela red">
+            </div>
+            <div class="row tabelasTimes justify-content-center">
+                <img src="tabela Vivo Keyd Stars.png" alt="tabela keyd">
+            </div>
+        </div>
+    </div>
     <script src="JavaScript/script.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js" integrity="sha384-I7E8VVD/ismYTF4hNIPjVp/Zjvgyol6VFvRkX/vR+Vc4jQkC+hVqc2pM8ODewa9r" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.min.js" integrity="sha384-0pUGZvbkm6XF6gxjEnlmuGrJXVbNuzT9qBBavbLwCsOGabYfZo0T0to5eqruptLy" crossorigin="anonymous"></script>
 </body>
+
 </html>
