+++ torneio/CBLOL/2024-1/cartinhas.py
